requirements.txt
Django==1.11
pytz==2017.3
django-registration-redux==2.0
